refactor(input): move sync_focused_webview to host, decouple ozma_webview from KeyboardFocused
Moves sync_focused_webview and its test cases from crates/ozma_webview into
src/input/focus.rs behind a new FocusSyncPlugin registered in the host binary.
ozma_webview no longer imports KeyboardFocused or OzmaTerminalInputSet.
