+++ crates/ozma_webview/src/control_plane.rs
@@ -1906,85 +1906,6 @@ mod focus_tests {
             "blur from the owning surface must clear focus"
         );
     }
-
-    #[test]
-    fn sync_preserves_app_declared_focus_from_control_plane() {
-        use crate::webview::render::sync_focused_webview;
-        use ozma_terminal::OzmaTerminal;
-
-        let mut app = bevy::app::App::new();
-        app.add_plugins(bevy::MinimalPlugins)
-            .init_resource::<FocusedWebview>()
-            .init_resource::<OzmaRegistry>()
-            .init_resource::<OzmuxRpc>()
-            .insert_resource(WebviewAssetRegistryRes(WebviewAssetRegistry::default()))
-            .add_systems(Update, sync_focused_webview);
-
-        // The owner surface is an OzmaTerminal (the terminal-level focus marker).
-        let surface = app.world_mut().spawn(OzmaTerminal).id();
-
-        app.world_mut().resource_mut::<OzmaRegistry>().insert(
-            "h1".into(),
-            OzmaView {
-                source: OzmaSource::Inline("<h1>x</h1>".into()),
-                entry: "index.html".into(),
-                interactive: true,
-                owner_surface: surface,
-                connection_id: 1,
-                forward_keys: vec![],
-                preload: vec![],
-            },
-        );
-        let child = app
-            .world_mut()
-            .spawn((
-                ChildOf(surface),
-                Webview {
-                    view_id: "h1".into(),
-                    instance_id: None,
-                    slot: 0,
-                },
-            ))
-            .id();
-
-        let (tx, rx) = unbounded::<ControlEvent>();
-        app.insert_resource(ControlEvents(rx));
-        tx.send(ControlEvent::SetFocus {
-            connection_id: 1,
-            owner_surface: surface,
-            handle: Some("h1".into()),
-            instance: None,
-        })
-        .unwrap();
-        app.world_mut()
-            .run_system_once(apply_control_events)
-            .unwrap();
-        assert_eq!(
-            app.world().resource::<FocusedWebview>().0,
-            Some(child),
-            "SetFocus must point FocusedWebview at the owned inline child"
-        );
-
-        // The regression: the per-frame sync must NOT clobber app-declared focus
-        // while the owning pane is live — the preserve arm covers it the
-        // same way it covers click-granted inline focus.
-        app.update();
-        assert_eq!(
-            app.world().resource::<FocusedWebview>().0,
-            Some(child),
-            "app-declared inline focus must survive the per-frame sync_focused_webview"
-        );
-
-        // Despawning the inline child fails the preserve arm, so the sync GC's
-        // it out of FocusedWebview.
-        app.world_mut().entity_mut(child).despawn();
-        app.update();
-        assert_eq!(
-            app.world().resource::<FocusedWebview>().0,
-            None,
-            "app-declared focus must clear once its inline child despawns"
-        );
-    }
 }
 
 #[cfg(test)]
